Updates for new build system.

+++ Frameworks/Syntrax-c/Syntrax-c/resampler.c
@@ -305,10 +305,10 @@ static int resampler_wrapper( resampler *r, sample_t out [], int* out_size,
 		sample_t const in [], int in_size )
 {
 	sample_t* out_ = out;
-	int result = resampler_inner_loop( r, &out_, out + *out_size, in, in_size ) - in;
+	long result = resampler_inner_loop( r, &out_, out + *out_size, in, in_size ) - in;
 
-	*out_size = out_ - out;
-	return result;
+	*out_size = (int)(out_ - out);
+	return (int) result;
 }
 
 static void resampler_fill( resampler *r )
